reuse existing gadgetry for detecting cycles
