+++ src/strict/eval.rs
@@ -5,7 +5,8 @@ use crate::finite_function::*;
 use crate::indexed_coproduct::*;
 use crate::semifinite::*;
 
-use crate::strict::layer::{converse, layer};
+use crate::strict::graph::converse;
+use crate::strict::layer::layer;
 use crate::strict::open_hypergraph::*;
 
 use num_traits::Zero;
