+++ src/strict/mod.rs
@@ -9,6 +9,7 @@ pub mod open_hypergraph;
 pub mod eval;
 pub mod functor;
 pub mod layer;
+pub mod graph;
 
 pub use crate::array::*;
 pub use crate::category::*;
