+++ src/strict/open_hypergraph/arrow.rs
@@ -1,4 +1,3 @@
-use crate::array::vec::VecKind;
 use crate::array::*;
 use crate::category::*;
 use crate::finite_function::*;
@@ -291,7 +290,11 @@ where
     }
 }
 
-impl<O, A> OpenHypergraph<VecKind, O, A> {
+impl<K: ArrayKind, O, A> OpenHypergraph<K, O, A>
+where
+    K::Type<K::I>: NaturalArray<K>,
+    K::Type<O>: Array<K, O>,
+{
     /// Returns true if there is no directed path from any node to itself.
     pub fn is_acyclic(&self) -> bool {
         self.h.is_acyclic()
